+++ packages/expo-sqlite/src/index.ts
@@ -1,4 +1,2 @@
-import { default as objects } from './SQLite';
-import * as types from './SQLite.types';
-const SQLite = {...objects, ... types};
-export { SQLite };  
+export * from './SQLite';
+export * from './SQLite.types';
