+++ packages/expo/src/globals.web.ts
@@ -9,9 +9,10 @@
 import Constants from 'expo-constants';
 import * as Font from 'expo-font';
 import * as Icon from '@expo/vector-icons';
+import * as SQLite from 'expo-sqlite';
 export { Asset } from 'expo-asset';
 export { Constants };
 export { Font };
 export { Icon };
 export { LinearGradient } from 'expo-linear-gradient';
-export { SQLite } from 'expo-sqlite';
+export { SQLite };
